Fix access rate policy bug causing incorrect initialization of water_mark_remote in access_rate_policy_dynamic.h

+++ cache_policy/access_rate_policy_dynamic.h
@@ -38,7 +38,7 @@ public:
     block_db_num_entries = block_cache_config.db.block_db.num_entries;
     cache_size = block_cache_config.cache.lru.cache_size;
     water_mark_local = 0.0;
-    water_mark_remote = (cache_size / block_db_num_entries);
+    water_mark_remote = float(cache_size / block_db_num_entries);
     water_mark_disk = 100.0;
     info("access_rate: {} and access_per_itr: {} and cache_size: {}", access_rate, access_per_itr, cache_size);
     info("water_mark_local: {} and water_mark_remote: {}", water_mark_local, water_mark_remote);
